+++ Samples~/ImportTester/Code/Types.cs
@@ -1,0 +1,10 @@
+﻿#nullable enable
+
+namespace At.Ac.FhStp.Import3D.ImportTester
+{
+    public enum LinearDirection
+    {
+        Next = 1,
+        Previous = -1
+    }
+}